refactor: toast component path
