+++ examples/basic/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './VideoTracksSelector';
 export * from './TextTracksSelector';
 export * from './Overlay';
 export * from './TopControl';
+export * from './Toast';
